requirements.txt
Django==5.2.5
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.4.0
django-allauth==0.60.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
python-decouple==3.8
whitenoise==6.8.0
gunicorn==21.2.0
